fix(serve): exclude browser voice from Windows builds
Browser voice drives pi through a PTY (github.com/creack/pty), which has
no Windows implementation: every call returns ErrUnsupported and the
Windows test job hangs in the serve/PTY path. Compile the feature out
there instead of shipping a voice session that connects and then cannot
touch the agent.

- voice.go, voice_agent.go, voice_gemini.go and their tests are now
  //go:build !windows; the mixed complexity_web_test.go is split so its
  voice half moves to complexity_voice_test.go under the same tag.
- voice_common.go (untagged) holds what both halves share: the JSON
  error shape and the pairing gate.
- voice_windows.go keeps the routes server.go registers and the
  EnableVoice entry point: /api/voice/config answers 200
  {enabled:false, reason:"voice is not supported on Windows"}, the
  session/ws endpoints answer 503 with the same reason, EnableVoice
  returns it as an error. voice_windows_test.go pins those shapes.
- internal/cli: enableServeVoice moves to serve_voice_unix.go; the
  Windows variant fails --voice at startup with a clear message before
  any key is resolved or listener opened. serve.go stays tag-free.
