+++ internal/webserver/voice.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 package webserver
 
 import (
@@ -242,30 +244,6 @@ func (s *ServerV2) handleDeleteVoiceSession(w http.ResponseWriter, r *http.Reque
 	}
 	s.voiceStore.delete(id)
 	w.WriteHeader(http.StatusNoContent)
-}
-
-// voiceHTTPError writes one error as JSON. The web UI reads these directly, so
-// the message is the user-facing text.
-func voiceHTTPError(w http.ResponseWriter, status int, err error) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-	_ = json.NewEncoder(w).Encode(map[string]any{"error": err.Error()})
-}
-
-// voiceAuthorized gates every voice endpoint behind the same pairing token the
-// terminal uses, answering 401 and reporting false when it is missing.
-//
-// This is not defense in depth, it is the actual boundary. `pi serve` binds all
-// interfaces, and a voice session can now type into the coding agent — which
-// can edit files and run commands. An unauthenticated caller reaching
-// POST /api/voice/sessions would therefore be reaching a shell, so voice must
-// be exactly as protected as the terminal it drives.
-func (s *ServerV2) voiceAuthorized(w http.ResponseWriter, r *http.Request) bool {
-	if s.pairingToken(r) {
-		return true
-	}
-	voiceHTTPError(w, http.StatusUnauthorized, fmt.Errorf("pair this browser with the server before using voice"))
-	return false
 }
 
 // EnableVoice configures the Gemini Live transport and verifies the key and
