+++ internal/webserver/voice_agent.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 package webserver
 
 import (
