+++ internal/webserver/voice_e2e_test.go
@@ -1,4 +1,4 @@
-//go:build e2e
+//go:build e2e && !windows
 
 package webserver
