+++ internal/webserver/voice_test.go
@@ -1,8 +1,9 @@
+//go:build !windows
+
 package webserver
 
 import (
 	"encoding/json"
-	"io"
 	"net/http"
 	"net/http/httptest"
 	"strings"
@@ -13,37 +14,6 @@ import (
 
 	"github.com/dimetron/pi-go/internal/voicegemini"
 )
-
-func testServer(t *testing.T) *ServerV2 {
-	t.Helper()
-	return NewServerV2(Config{Addr: "127.0.0.1:0"})
-}
-
-// voiceToken pairs a browser with s and returns the approved token. Voice
-// endpoints are gated on it, so every test that reaches one has to hold it —
-// which is the point: a voice session can type into the coding agent.
-func voiceToken(t *testing.T, s *ServerV2) string {
-	t.Helper()
-	code, _, err := s.BootstrapPair(".")
-	if err != nil {
-		t.Fatalf("BootstrapPair() = %v", err)
-	}
-	token, err := s.pairingMgr.Approve(code)
-	if err != nil {
-		t.Fatalf("Approve() = %v", err)
-	}
-	return token
-}
-
-// voiceReq builds a request to a voice endpoint carrying an approved token.
-func voiceReq(t *testing.T, s *ServerV2, method, target string, body io.Reader) *http.Request {
-	t.Helper()
-	sep := "?"
-	if strings.Contains(target, "?") {
-		sep = "&"
-	}
-	return httptest.NewRequest(method, target+sep+"token="+voiceToken(t, s), body)
-}
 
 // withVoice returns a server whose voice transport is configured without
 // touching the network — EnableVoice would verify against Google.
